+++ src/witsec-modal-window-block/component.js
@@ -1,8 +1,9 @@
 mbrApp.loadComponents(
 	"witsec-modal-window",
 	{"witsec-modal-window-block":{
-			_group:"Article",
+			_group:"witsec",
 			_onParamsShow: function(e,$params,$block) {
+				// Replace all characters (except letters, numbers and dashes) to dashes, so modal names end up with proper names
 				this._params.modalName = this._params.modalName.replace(/([^a-z0-9\-]+)/img, '-');
 
 				// If we don't encode the textareas, it'll mess up the HTMLCode textarea in the gear box
@@ -13,9 +14,11 @@ mbrApp.loadComponents(
 			_params:{
 				modalNotice:         {type:"separator",title: "The block on the left will not be visible on preview/publish."},
 				modalName:           {type:"text",title:"Modal Name (must be unique)",default:""},
+				modalSize:           {type:"select",title:"Size",values:{small:"Small",medium:"Medium",large:"Large"},default:"medium"},
 				modalHeader:         {type:"switch",title:"Header",default:!0,condition:["modalName"]},
 				modalTitle:          {type:"text",title:"Title",default:"Modal title",condition:["modalName", "modalHeader"]},
 				modalBody:           {type:"textarea",title:"Body",default:"Modal body text goes here.",condition:["modalName"]},
+				modalBodyHTML:       {type:"switch",title:"Enable HTML",default:!1,condition:["modalName"]},
 				modalFooter:         {type:"switch",title:"Footer",default:!0,condition:["modalName"]},
 				modalCloseBtn:       {type:"switch",title:"Close Button",default:!0,condition:["modalName", "modalFooter"]},
 				modalCloseText:      {type:"text",title:"Close Button Caption",default:"Close",condition:["modalName", "modalFooter", "modalCloseBtn"]},
@@ -24,7 +27,7 @@ mbrApp.loadComponents(
 				modalLinkText:       {type:"text",title:"Link Text",default:"Visit Mobirise",condition:["modalName", "modalFooter", "modalLink"]},
 				modalLinkNewWindow:  {type:"switch",title:"Open in New Window",default:!0,condition:["modalName", "modalFooter", "modalLink"]},
 				modalVerticalCenter: {type:"switch",title:"Vertically Centered",default:!1,condition:["modalName"]},
-				modalFade:           {type:"switch",title:"Fade",default:!0,condition:["modalName"]}
+				modalFade:           {type:"switch",title:"Fade Effect",default:!0,condition:["modalName"]}
 			},
 			modalTest: "Please use the gear-icon to change the modal name.",
 			modalBody: "Modal body text goes here.",
@@ -40,22 +43,42 @@ mbrApp.loadComponents(
 						this.modalWindow = "";
 					}
 					else {
-						this.modalTest = '<div class="mbr-section-btn"><a href="#" class="btn btn-primary display-4" data-toggle="modal" data-target="#' + val + '">Preview ' + val + '</a></div>';
+						// Buttons to preview modal and how-to-use window
+						this.modalTest  = '<div class="mbr-section-btn">';
+						this.modalTest += '<a href="#" class="btn btn-primary display-4" data-toggle="modal" data-target="#' + val + '">Preview ' + val + '</a>';
+						this.modalTest += '<a href="#" class="btn btn-primary display-4" data-toggle="modal" data-target="#' + val + '-howtouse">How to use ' + val + '</a>';
+						this.modalTest += '</div>';
+
+						// Modal window for how-to-use
+						this.modalTest += '<div class="modal fade" id="' + val + '-howtouse" tabindex="-1" role="dialog" aria-labelledby="' + val + '-howtouseLabel" aria-hidden="true"><div class="modal-dialog" role="document"><div class="modal-content"><div class="modal-header"><h5 class="modal-title" id="' + val + '-howtouseLabel">How to use?</h5><a href="" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></a></div>';
+						this.modalTest += '<div class="modal-body">Using your new modal is very easy. Simply create a new link, then click the "..." tab and enter the following as custom URL:<br /><br /><code>javascript:OpenModal(\'' + val + '\')</code></div>';
+						this.modalTest += '<div class="modal-footer"><div class="mbr-section-btn"><a href="#" class="btn btn-secondary display-4" data-dismiss="modal">Close</a></div></div></div></div></div>';
 					}
 				}
 
 				if (param == "modalBody") {
-					// A little hack regarding textareas, in case someone would ever want to add a textarea to this type of modal window...
+					// We write the modalBody to the "html"
 					this.modalBody = val;
+					
+					// A little hack regarding textareas, in case someone would ever want to add a textarea to this type of modal window...
 					this._params.modalBody = this.modalBody.replace(/<\/textarea/gim, "&lt;/textarea");
 				}
 
 				// No matter what param changes, we need to change the modal to reflect this change
 				if (this._params.modalName != "") {
 					var p = this._params;
+					
+					switch (p.modalSize) {
+						case "small":	var size = "modal-sm";
+										break;
+						case "large":	var size = "modal-lg";
+										break;
+						default:		var size = "";
+					}
+
 					var m = "";
 					m += '<div class="modal ' + (p.modalFade ? "fade" : "") + '" id="' + p.modalName + '" tabindex="-1" role="dialog" aria-labelledby="' + p.modalName + 'Label" aria-hidden="true">';
-					m += '  <div class="modal-dialog ' + (p.modalVerticalCenter ? "modal-dialog-centered" : "") + '" role="document">';
+					m += '  <div class="modal-dialog ' + size + ' ' + (p.modalVerticalCenter ? "modal-dialog-centered" : "") + '" role="document">';
 					m += '    <div class="modal-content">';
 
 					// Header
@@ -67,7 +90,15 @@ mbrApp.loadComponents(
 					}
 
 					// Body
-					m += '<div class="modal-body">' + this.modalBody + '</div>';
+					if (p.modalBodyHTML) {
+						// If HTML is enabled, accept everything
+						b = this.modalBody;
+					}
+					else {
+						// If HTML is disabled, replace any "<" with "&lt;" and replace \n with HTML breaklines
+						b = this.modalBody.replace(/</g, "&lt;").replace(/\n|\r/g, "<br />");
+					}
+					m += '<div class="modal-body">' + b + '</div>';
 
 					// Footer
 					if (p.modalFooter) {