+++ CHANGELOG.md
@@ -2,17 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.0 (2018-11-12)
+- Added 'Enable HTML' switch in params menu. When enabled, the user can enter all the HTML (s)he likes. When disabled, breaklines (`\n`) get replaced with `<br />`. This option is turned off by default, so if HTML is used with a previous version, this switch has to be turned on in order for it to work properly
+- Added 'Size' in params menu. This gives the user the option to change the size of the modal window (small/medium/large)
+- Added 'How to use' button that explains the user how to link to the newly created modal window. Like the 'Preview' button, this will be removed on Preview/Publish
+- Added 'Learn more...' link in extension description
+- Changed group to 'witsec'
+
 ## 0.3 (2018-11-05)
 
-### Fixed
 - Fixed modal window contents, so they don't animate (fade up) anymore, regardless of site settings
 
 ## 0.2 (2018-11-05)
 
-### Fixed
 - Fixed block's thumb, showing incorrect extension
 
 ## 0.1 (2018-11-05)
 
-### Summary
 This is the initial release of the Modal Window extension for Mobirise.